refactor(live-common/device): move fetchMcusUseCase out of subfolder
