+++ libs/ledger-live-common/src/manager/api.ts
@@ -46,7 +46,7 @@ import {
   secureChannelMock,
 } from "../socket/socket.mock";
 import { getProviderId } from "./provider";
-import fetchMcusUseCase from "../device/use-cases/managerApi/fetchMcusUseCase";
+import fetchMcusUseCase from "../device/use-cases/fetchMcusUseCase";
 
 declare global {
   namespace NodeJS {
